Move OVMS C-API calculator to tests
