+++ src/test/mediapipe/calculators/exceptionduringclosecalculator.cpp
@@ -25,7 +25,7 @@
 #include "mediapipe/framework/calculator_framework.h"
 #include "mediapipe/framework/port/canonical_errors.h"
 #pragma GCC diagnostic pop
-#include "src/mediapipe_calculators/ovmscalculator.pb.h"
+#include "src/test/mediapipe/calculators/ovmscalculator.pb.h"
 // here we need to decide if we have several calculators (1 for OVMS repository, 1-N inside mediapipe)
 // for the one inside OVMS repo it makes sense to reuse code from ovms lib
 namespace mediapipe {
