+++ src/test/mediapipe/calculators/ovms_kfs_calculator.cc
@@ -13,27 +13,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //*****************************************************************************
+#include "../../../stringutils.hpp"
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wdeprecated-declarations"
 #include "mediapipe/framework/calculator_framework.h"
-#include "mediapipe/framework/formats/image_frame.h"
 #include "mediapipe/framework/port/canonical_errors.h"
-#include "src/kfs_frontend/kfs_grpc_inference_service.hpp"
-#include "src/mediapipe_calculators/ovmscalculator.pb.h"
 #pragma GCC diagnostic pop
+#include "src/kfs_frontend/kfs_grpc_inference_service.hpp"
+#include "src/test/mediapipe/calculators/ovmscalculator.pb.h"
 
+// here we need to decide if we have several calculators (1 for OVMS repository, 1-N inside mediapipe)
+// for the one inside OVMS repo it makes sense to reuse code from ovms lib
 namespace mediapipe {
 
 using std::endl;
 
-class OVMSTestImageInputPassthroughCalculator : public CalculatorBase {
+class OVMSTestKFSPassCalculator : public CalculatorBase {
+    std::shared_ptr<KFSResponse> response;
+
 public:
     static absl::Status GetContract(CalculatorContract* cc) {
         RET_CHECK(!cc->Inputs().GetTags().empty());
         RET_CHECK(!cc->Outputs().GetTags().empty());
         RET_CHECK(cc->Inputs().GetTags().size() == 1);
-        cc->Inputs().Tag("IMAGE").Set<ImageFrame>();
-        cc->Outputs().Tag("IMAGE").Set<ImageFrame>();
+        RET_CHECK(cc->Outputs().GetTags().size() == 1);
+        cc->Inputs().Tag("REQUEST").Set<const KFSRequest*>();
+        cc->Outputs().Tag("RESPONSE").Set<KFSResponse*>();
 
         return absl::OkStatus();
     }
@@ -52,15 +57,26 @@ public:
     }
 
     absl::Status Process(CalculatorContext* cc) final {
-        LOG(INFO) << "Process start";
-        const auto& received = cc->Inputs().Tag("IMAGE").Get<ImageFrame>();
-        ImageFrame image = ImageFrame(received.Format(), received.Width(), received.Height());
-        image.CopyFrom(received, 1);
-        cc->Outputs().Tag("IMAGE").AddPacket(MakePacket<ImageFrame>(std::move(image)).At(Timestamp(0)));
+        const KFSRequest* request = cc->Inputs().Tag("REQUEST").Get<const KFSRequest*>();
+        response = std::make_shared<KFSResponse>();
+        for (int i = 0; i < request->inputs().size(); i++) {
+            auto* output = response->add_outputs();
+            output->set_datatype(request->inputs()[i].datatype());
+            output->set_name("out");
+            for (int j = 0; j < request->inputs()[i].shape_size(); j++) {
+                output->add_shape(request->inputs()[i].shape().at(j));
+            }
+        }
+
+        for (int i = 0; i < request->raw_input_contents().size(); i++) {
+            response->add_raw_output_contents()->assign(request->raw_input_contents()[i].data(), request->raw_input_contents()[i].size());
+        }
+
+        cc->Outputs().Tag("RESPONSE").AddPacket(::mediapipe::MakePacket<KFSResponse*>(response.get()).At(cc->InputTimestamp()));
 
         return absl::OkStatus();
     }
 };
 
-REGISTER_CALCULATOR(OVMSTestImageInputPassthroughCalculator);
+REGISTER_CALCULATOR(OVMSTestKFSPassCalculator);
 }  // namespace mediapipe
